+++ test/runtests.jl
@@ -34,6 +34,7 @@ for file in [
     "test_remark.jl",
     "test_rounds.jl",
     "test_rounds_papers.jl",
+    "test_jsq.jl",
     "test_quality.jl",
 ]
     path = joinpath(@__DIR__, file)